Implement NFA simulation and fix `QuantifierKind::is_satisfied` implementation
Previously I tried to implement `fsm::nfa::Run` by keeping a reference
to the state in the NFA. This did not work with the rust borrow checker
as it does not know I only use the NFA as non-mutable references. This
could be solved by wrapping most fields in the simulator as `std::cell::RefCell`'s, but would be ugly and require runtime verification of the borrow rules. I instead opted to keep track of the states using their ids. This also allows me to hash a run, which is needed for the `fsm::nfa::NfaSimulator::eps_closure` function, which checks the eps_closure based on runtime information, like state counters.