Fix run simulation for nested groups and quantifiers
The first idea was to count how many times we visited a state and use
that information to take quantified transitions. This however is
incorrect when using nested quantifiers/groups. To solve this, we now
reset the visited count for a state when the quantified transition is
taken for that run of the NFA.