Implement `ItemSet` closure and add `Epsilon` as a grammar symbol
The `Epsilon` symbol is ignored by the cursor of the `ItemBody`. When
initializing, it sets the cursor to one symbol beyond a consecutive
chain of epsilon symbols when it is the first symbol of the rule. When
advancing the cursor, it will keep advancing until a non-epsilon symbol
or the end of the rule.